// crates/rupu-cp/web/src/lib/codenamePalette.gen.ts
export type BadgeShape =
  | 'circle'
  | 'triangle'
  | 'square'
  | 'diamond'
  | 'pentagon'
  | 'hexagon'
  | 'star'
  | 'cross'
  | 'ring'
  | 'chevron'
  | 'inv-triangle'
  | 'half';